refactor: separate env name prompts
