+++ packages/amplify-cli/src/prompts/index.ts
@@ -1,5 +1,6 @@
 export * from './default-editor';
-export * from './env-name';
+export * from './env-name-input';
+export * from './env-name-select';
 export * from './frontend-handler';
 export * from './project-name';
 export * from './providers';
